MC-84621: change type of paymentSum

+++ src/main/java/ru/moysklad/remap_1_2/entities/documents/GiftCard.java
@@ -6,13 +6,11 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import ru.moysklad.remap_1_2.entities.MetaEntity;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class GiftCard extends MetaEntity {
     private String name;
-    private BigDecimal paymentSum;
+    private Long paymentSum;
 }
